+++ src/class-generators.js
@@ -2,16 +2,12 @@ const vscode = require('vscode');
 const miscLib = require('./misc-lib')
 
 function createGetterAndSetter(propsText) {
-    var properties = propsText.split('\r\n').filter(x => x.length > 2);
+    var properties = propsText.split('\r\n').filter(x => {
+        return x.length > 2
+    });
     var propList = miscLib.extractPropertiesArray(properties);
 
-    var generatedCode = 
-`
-
-'***************************
-'*** GETTERS AND SETTERS ***
-'***************************
-`;
+    var generatedCode = `\r\n\r\n'***************************\r\n'*** GETTERS AND SETTERS ***\r\n'***************************\r\n`;
     var errorMessage = `Something went wrong (fix them before proceeding) in the following lines: `;
     var errorList = propList.filter(p => !p.createStatus);
 
@@ -27,25 +23,11 @@ function createGetterAndSetter(propsText) {
                 // Removal of variant arrays parenthesis
                 if (rawAttribute.indexOf("()") !== -1) rawAttribute = rawAttribute.replace("()", "");
                 if (p.attribute.indexOf("()") !== -1) p.attribute = p.attribute.replace("()", "");
-                let code = 
-`
-''''''''''''''''''''''
-' ${miscLib.transformFirstCharToUpperCase(rawAttribute)} property
-''''''''''''''''''''''
-Public Property Get ${rawAttribute}() As ${p.type}
-\t${rawAttribute} = ${p.attribute}
-End Property
-`;
+                let code = `\r\n''''''''''''''''''''''\r\n' ${miscLib.transformFirstCharToUpperCase(rawAttribute)} property\r\n''''''''''''''''''''''\r\nPublic Property Get ${rawAttribute}() As ${p.type}\r\n\t${rawAttribute} = ${p.attribute}\r\nEnd Property\r\n`;
                 generatedCode += code;
     
                 if (!p.constStatus) {
-                    let code = 
-`
-Public Property Let ${rawAttribute}(value As ${p.type})
-\t${p.attribute} = value
-End Property
-`;
-
+                    let code = `\r\nPublic Property Let ${rawAttribute}(value As ${p.type})\r\n\t${p.attribute} = value\r\nEnd Property\r\n`;
                     generatedCode += code;
                 }
             }
@@ -62,13 +44,7 @@ function createConstructor(propsText) {
     var properties = propsText.split('\r\n').filter(x => x.length > 2);
     var propList = miscLib.extractPropertiesArray(properties);
 
-    var generatedCode = 
-`
-
-'************************
-'*** MAIN CONSTRUCTOR ***
-'************************
-`;
+    var generatedCode = `\r\n\r\n'************************\r\n'*** MAIN CONSTRUCTOR ***\r\n'************************\r\n`;
 
     var errorMessage = `Something went wrong (fix them before proceeding) in the following lines: `;
     var errorList = propList.filter(p => !p.createStatus);
@@ -79,9 +55,7 @@ function createConstructor(propsText) {
         else vscode.window.showErrorMessage(errorMessage + errObjMsg.innerErrorsLines);
     }
     else {
-        let codeSignature = 
-`
-Public Sub Init(`;
+        let codeSignature = `\r\nPublic Sub Init(`;
 
         for (var i = 0, j = 0; i < propList.length; i++, j++) {
             if (!propList[i].constStatus) {
@@ -95,9 +69,7 @@ Public Sub Init(`;
                             j = 0;
                             // Don't break line if the next el is the last and is constant
                             if (i !== propList.length - 2 && !propList[i+1].constStatus) {
-                                codeSignature +=
-` _ 
-`;
+                                codeSignature += ` _ \r\n`;
                             }
                         }
                     }
@@ -115,11 +87,9 @@ Public Sub Init(`;
             }
         }
 
-        codeSignature += `)
-`;
+        codeSignature += `)\r\n`;
         
-        let codeBody =
-``;
+        let codeBody = ``;
         for (let p of propList) {
             if (!p.constStatus) {
                 let rawAttribute = p.attribute.split('_')[1]
@@ -127,26 +97,14 @@ Public Sub Init(`;
                     let codeInnerBody = ``;
     
                     if (p.type == 'Integer' || p.type == 'Double' || p.type == 'Long') {
-                        codeInnerBody = 
-`\tIf ${p.attribute} = 0 Then
-\t\t${p.attribute} = a${miscLib.transformFirstCharToUpperCase(rawAttribute)}
-\tEnd If
-`;
+                        codeInnerBody = `\tIf ${p.attribute} = 0 Then\r\n\t\t${p.attribute} = a${miscLib.transformFirstCharToUpperCase(rawAttribute)}\r\n\tEnd If\r\n`;
                     }
                     else if (p.type == 'String') {
-                        codeInnerBody = 
-`\tIf ${p.attribute} = "" Then
-\t\t${p.attribute} = a${miscLib.transformFirstCharToUpperCase(rawAttribute)}
-\tEnd If
-`;
+                        codeInnerBody = `\tIf ${p.attribute} = "" Then\r\n\t\t${p.attribute} = a${miscLib.transformFirstCharToUpperCase(rawAttribute)}\r\n\tEnd If\r\n`;
                     }
                     // Variant types
                     else {
-                        codeInnerBody = 
-`\tIf isEmpty(${p.attribute}) Then
-\t\t${p.attribute} = a${miscLib.transformFirstCharToUpperCase(rawAttribute)}
-\tEnd If
-`;
+                        codeInnerBody = `\tIf isEmpty(${p.attribute}) Then\r\n\t\t${p.attribute} = a${miscLib.transformFirstCharToUpperCase(rawAttribute)}\r\n\tEnd If\r\n`;
                     }
                     
                     codeBody += codeInnerBody;
@@ -158,8 +116,7 @@ Public Sub Init(`;
             }
         }
 
-        let codeFooter = 
-`End Sub`;
+        let codeFooter = `End Sub`;
         generatedCode += codeSignature + codeBody + codeFooter;
             
     }
@@ -171,9 +128,9 @@ function createAttributesList(propsText) {
     var properties = propsText.split('\r\n').filter(x => x.length > 2);
     var propList = miscLib.extractPropertiesArray(properties);
 
-    var generatedCode = 
-`
-`;
+    var attributesProperty = `\r\nPrivate p_attributesList() As Variant\r\n`;
+
+    var generatedCode = `\r\n`;
 
     var errorMessage = `Something went wrong (fix them before proceeding) in the following lines: `;
     var errorList = propList.filter(p => !p.createStatus);
@@ -184,15 +141,7 @@ function createAttributesList(propsText) {
         else vscode.window.showErrorMessage(errorMessage + errObjMsg.innerErrorsLines);
     }
     else {
-        generatedCode += 
-`Private p_attributesList() As Variant
-
-'***************************************************
-'*** CLASS ARRAY LIST INITIALIZATION WITH VALUES ***
-'***************************************************
-
-Private Sub class_initialize()
-\tp_attributesList = Array(`;
+        generatedCode += `\r\n'***************************************************\r\n'*** CLASS ARRAY LIST INITIALIZATION WITH VALUES ***\r\n'***************************************************\r\nPrivate Sub class_initialize()\r\n\tp_attributesList = Array(`;
 
         for (let i = 0, j = 0; i < propList.length; i++, j++) {
             let rawAttribute = propList[i].attribute.split('_')[1]
@@ -203,14 +152,11 @@ Private Sub class_initialize()
                     // Breaking lines (VB editor in excel has character limits)
                     if (j == 5) {
                         j = 0;
-                        generatedCode +=
-` _ 
-`;
+                        generatedCode += ` _ \r\n`;
                     }
                 }
                 else {
-                    generatedCode += `)
-End Sub`;
+                    generatedCode += `)\r\nEnd Sub`;
                 }
             }
             else {
@@ -219,16 +165,16 @@ End Sub`;
         }
     }
 
-    return generatedCode;
+    return { generatedCode, attributesProperty };
 }
 
 function createAttributesWithFormatList(propsText) {
     var properties = propsText.split('\r\n').filter(x => x.length > 2);
     var propList = miscLib.extractPropertiesArray(properties);
 
-    var generatedCode = 
-`
-`;
+    var attributesProperty = `\r\nPrivate p_attributesList() As Variant\r\nPrivate p_attributesFormatTypesList() As Variant\r\nPrivate p_attributesFormatValuesList() As Variant\r\n`;
+
+    var generatedCode = `\r\n`;
 
     var errorMessage = `Something went wrong (fix them before proceeding) in the following lines: `;
     var errorList = propList.filter(p => !p.createStatus);
@@ -239,17 +185,7 @@ function createAttributesWithFormatList(propsText) {
         else vscode.window.showErrorMessage(errorMessage + errObjMsg.innerErrorsLines);
     }
     else {
-        generatedCode += 
-`Private p_attributesList() As Variant
-Private p_attributesFormatTypesList() As Variant
-Private p_attributesFormatValuesList() As Variant
-
-'***************************************************
-'*** CLASS ARRAY LIST INITIALIZATION WITH VALUES ***
-'***************************************************
-
-Private Sub class_initialize()
-\tp_attributesList = Array(`;
+        generatedCode += `\r\n'***************************************************\r\n'*** CLASS ARRAY LIST INITIALIZATION WITH VALUES ***\r\n'***************************************************\r\nPrivate Sub class_initialize()\r\n\tp_attributesList = Array(`;
 
         for (let i = 0, j = 0; i < propList.length; i++, j++) {
             let rawAttribute = propList[i].attribute.split('_')[1]
@@ -260,9 +196,7 @@ Private Sub class_initialize()
                     // Breaking lines (VB editor in excel has character limits)
                     if (j == 5) {
                         j = 0;
-                        generatedCode +=
-` _ 
-`;
+                        generatedCode += ` _ \r\n`;
                     }
                 }
                 else {
@@ -271,9 +205,7 @@ Private Sub class_initialize()
             }
         }
 
-        generatedCode += 
-`
-\tp_attributesFormatTypesList = Array(`;
+        generatedCode += `\r\n\tp_attributesFormatTypesList = Array(`;
 
         for (let i = 0, j = 0; i < propList.length; i++, j++) {
             generatedCode += `"${propList[i].formatType}"`;
@@ -282,9 +214,7 @@ Private Sub class_initialize()
                 // Breaking lines (VB editor in excel has character limits)
                 if (j == 5) {
                     j = 0;
-                    generatedCode +=
-` _ 
-`;
+                    generatedCode += ` _ \r\n`;
                 }
             }
             else {
@@ -292,9 +222,7 @@ Private Sub class_initialize()
             }
         }
 
-        generatedCode += 
-`
-\tp_attributesFormatValuesList = Array(`;
+        generatedCode += `\r\n\tp_attributesFormatValuesList = Array(`;
 
         for (let i = 0, j = 0; i < propList.length; i++, j++) {
             generatedCode += `"${propList[i].formatValue}"`;
@@ -303,19 +231,16 @@ Private Sub class_initialize()
                 // Breaking lines (VB editor in excel has character limits)
                 if (j == 5) {
                     j = 0;
-                    generatedCode +=
-` _ 
-`;
+                    generatedCode += ` _ \r\n`;
                 }
             }
             else {
-                generatedCode += `)
-End Sub`;
+                generatedCode += `)\r\nEnd Sub`;
             }
         }
     }
 
-    return generatedCode;
+    return { generatedCode, attributesProperty };
 }
 
 module.exports = {