+++ CHANGELOG.md
@@ -36,3 +36,7 @@
 ## v1.2.2
 - Bug fixes: Use of regExp to check sentence structure + error handling in a more concise way
 - README updated with more detailed usage description
+
+## v1.3.0
+- Bug fixes: Changed breaking lines method in the class generators and regex evaluation cases for constant properties with format
+- Features: Full class generator with or without factory